added a menu list to the landlord available room to be able to edit or delete and also added a conditional statement to render user menu base on role

+++ backend/contollers/moderatorController.js
@@ -53,12 +53,13 @@ const updateRoom = async(req, res) => {
 
 const deleteRoom = async(req, res) => {
   const { roomId } = req.params
+  const userId = req.user._id
   try {
     const room = await Room.findById(roomId)
     if (!room) {
       return res.status(404).json({ message: 'Room not found' });
     }
-
+    if(room.landlord.toString() !== userId.toString()) return res.status(403).json({message: 'You are not authorized to delete this room'})
     await Room.deleteOne({_id: roomId})
     res.status(200).json({message: 'Room deleted successful'})
   } catch (error) {
